Create script.js
Collected name, age, and occupation from the user and stored them in an object. 
Implemented functionality to add and list products in the shopping cart. 
Calculated the total price of the cart using the reduce() method. 
Enabled dynamic product addition from the user. 
Researched and implemented a function to remove products from the cart.